requirements.txt
psycopg2
gunicorn
cchardet
python-memcached
awesome-slugify
regex
datadog
flask-cors
flask-jsonpify
sqlalchemy
six
redis
git+git://github.com/openspending/babbage.fiscal-data-package.git
git+git://github.com/openspending/babbage.git@feature/optimize-member-queries#egg=babbage==0.2.0
git+git://github.com/frictionlessdata/datapackage-py.git@v0.8.4#egg=datapacakge==0.8.4
git+git://github.com/frictionlessdata/jsontableschema-py.git@v0.8.3#egg=jsontableschema==0.8.3
git+git://github.com/frictionlessdata/jsontableschema-sql-py.git#egg=jsontableschema-sql==0.4.0
git+git://github.com/akariv/tabulator-py.git@feature/fake-user-agent#egg=tabulator==0.11.1
